requirements.txt
praat-parselmouth
tqdm
joblib
click
